+++ CHANGELOG.md
@@ -14,6 +14,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add method `AsyncMonitoredItem::delete()` to release server resources in a controlled manner
   asynchronously before dropping.
 - Add `MonitoredItemCreateRequestBuilder`.
+- `AsyncClient` and `AsyncSubscription` no longer depend on the `tokio` feature.
 - Experimental: New function `create_monitored_items_callback()` with accompanying type
   `MonitoredItemHandle`. Hidden behind the optional feature `experimental-monitored-item-callback`
   for stabilizing the public API.